ADD NormalFoodRecord API + Modified AI-FoodRecord API
