+++ src/main/java/com/springboot/peanut/controller/FoodController.java
@@ -5,12 +5,12 @@ import com.springboot.peanut.dto.food.FoodNutritionDto;
 import com.springboot.peanut.dto.food.FoodPredictResponseDto;
 import com.springboot.peanut.dto.signDto.ResultDto;
 import com.springboot.peanut.service.Food.FoodAIService;
-import com.springboot.peanut.service.Food.FoodRecordAIService;
 import com.springboot.peanut.service.Food.FoodRecordNormalService;
 import io.swagger.annotations.ApiImplicitParam;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.method.P;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,16 +23,15 @@ import java.util.List;
 @RequiredArgsConstructor
 public class FoodController {
     private final FoodAIService foodAIService;
-    private final FoodRecordAIService foodRecordAIService;
     private final FoodRecordNormalService foodRecordNormalService;
 
 
     @PostMapping("/ai/predict")
     @ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 발급 받은 access_token", required = true, dataType = "String", paramType = "header")
     public ResponseEntity<FoodPredictResponseDto> BlockPlay(
-            @RequestPart("foodImage") MultipartFile foodImage) throws IOException {
+            @RequestPart("foodImage") MultipartFile foodImage,HttpServletRequest request) throws IOException {
 
-        FoodPredictResponseDto results = foodAIService.FoodNamePredict(foodImage);
+        FoodPredictResponseDto results = foodAIService.FoodNamePredict(foodImage,request);
         return ResponseEntity.status(HttpStatus.OK).body(results);
 
     }
@@ -40,14 +39,21 @@ public class FoodController {
     @GetMapping("/ai/details")
     @ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 발급 받은 access_token", required = true, dataType = "String", paramType = "header")
     public ResponseEntity<List<FoodDetailInfoDto>> getFoodDetailInfo(@RequestParam  List<String> name, HttpServletRequest request){
-        List<FoodDetailInfoDto> foodDetailInfoDto = foodRecordAIService.getFoodDetailInfo(name,request);
+        List<FoodDetailInfoDto> foodDetailInfoDto = foodAIService.getFoodDetailInfo(name,request);
         return ResponseEntity.status(HttpStatus.OK).body(foodDetailInfoDto);
     }
 
     @PostMapping("/ai/save-meal")
     @ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 발급 받은 access_token", required = true, dataType = "String", paramType = "header")
     public ResponseEntity<ResultDto> createAIMealInfo(String mealTime, HttpServletRequest request){
-        ResultDto resultDto = foodRecordAIService.createAIMealInfo(mealTime,request);
+        ResultDto resultDto = foodAIService.createAIMealInfo(mealTime,request);
+        return ResponseEntity.status(HttpStatus.OK).body(resultDto);
+    }
+
+    @PostMapping("/normal/save-image")
+    @ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 발급 받은 access_token", required = true, dataType = "String", paramType = "header")
+    public ResponseEntity<ResultDto> saveNormalMealInfoImage(@RequestPart("foodImage") MultipartFile foodImage, HttpServletRequest request) throws IOException {
+        ResultDto resultDto = foodRecordNormalService.saveNormalMealInfoImage(foodImage,request);
         return ResponseEntity.status(HttpStatus.OK).body(resultDto);
     }
