+++ src/main/java/com/springboot/peanut/dto/food/FoodCheckDto.java
@@ -1,9 +1,13 @@
 package com.springboot.peanut.dto.food;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
 import java.util.List;
 
+@Data
+@AllArgsConstructor
 public class FoodCheckDto {
-    private String imageUrl;
 
     private List<String> foodName;
