+++ src/main/java/com/springboot/peanut/service/Food/Impl/FoodAIServiceImpl.java
@@ -2,10 +2,17 @@ package com.springboot.peanut.service.Food.Impl;
 
 import com.springboot.peanut.S3.S3Uploader;
 import com.springboot.peanut.dao.FoodPredictDao;
+import com.springboot.peanut.dao.MealDao;
+import com.springboot.peanut.dto.food.FoodDetailInfoDto;
 import com.springboot.peanut.dto.food.FoodPredictDto;
 import com.springboot.peanut.dto.food.FoodPredictResponseDto;
-import com.springboot.peanut.entity.FoodPredict;
+import com.springboot.peanut.dto.signDto.ResultDto;
+import com.springboot.peanut.entity.*;
+import com.springboot.peanut.repository.BloodSugar.BloodSugarRepository;
+import com.springboot.peanut.repository.FoodNutrition.FoodNutritionRepository;
 import com.springboot.peanut.service.Food.FoodAIService;
+import com.springboot.peanut.service.Jwt.JwtAuthenticationService;
+import com.springboot.peanut.service.Result.ResultStatusService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,9 +23,12 @@ import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientException;
 import reactor.core.publisher.Mono;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,15 +39,26 @@ public class FoodAIServiceImpl implements FoodAIService {
     private final S3Uploader s3Uploader;
     private final WebClient webClient;
     private final Logger logger = LoggerFactory.getLogger(FoodAIService.class);
+    private final FoodNutritionRepository foodNutritionRepository;
+    private final BloodSugarRepository bloodSugarRepository;
+    private final MealDao mealDao;
+    private final JwtAuthenticationService jwtAuthenticationService;
+    private final ResultStatusService resultStatusService;
 
-    public FoodAIServiceImpl(FoodPredictDao foodPredictDao, S3Uploader s3Uploader, WebClient.Builder webClientBuilder) throws IOException {
+    public FoodAIServiceImpl(FoodPredictDao foodPredictDao, S3Uploader s3Uploader, WebClient.Builder webClientBuilder,FoodNutritionRepository foodNutritionRepository,
+                             BloodSugarRepository bloodSugarRepository,MealDao mealDao,JwtAuthenticationService jwtAuthenticationService,ResultStatusService resultStatusService  ) throws IOException {
         this.foodPredictDao = foodPredictDao;
         this.s3Uploader = s3Uploader;
         this.webClient = webClientBuilder.baseUrl("http://52.78.145.50:8000").build();
+        this.foodNutritionRepository = foodNutritionRepository;
+        this.bloodSugarRepository = bloodSugarRepository;
+        this.mealDao = mealDao;
+        this.jwtAuthenticationService = jwtAuthenticationService;
+        this.resultStatusService = resultStatusService;
     }
 
     @Override
-    public FoodPredictResponseDto FoodNamePredict(MultipartFile foodImage) {
+    public FoodPredictResponseDto FoodNamePredict(MultipartFile foodImage,HttpServletRequest request) {
         FoodPredictResponseDto foodPredictResponseDto = null;
         try {
             String imageUrl = s3Uploader.uploadImage(foodImage, "peanut/before");
@@ -52,6 +73,7 @@ public class FoodAIServiceImpl implements FoodAIService {
                     .bodyToMono(FoodPredictResponseDto.class);
 
             foodPredictResponseDto = response.block();
+
             logger.info("Received Response: {}", foodPredictResponseDto);
 
             if (foodPredictResponseDto != null) {
@@ -62,8 +84,9 @@ public class FoodAIServiceImpl implements FoodAIService {
                 foodPredictDao.saveFoodPredictResults(foodPredict);
 
                 logger.info("Saved FoodPredict Result: {}", foodPredict);
-            }
 
+            }
+            request.getSession().setAttribute("imageUrl", foodPredictResponseDto.getImage_url());
         } catch (WebClientException | IOException e) {
             logger.error("Error occurred while fetching results: ", e);
             e.printStackTrace();
@@ -71,4 +94,98 @@ public class FoodAIServiceImpl implements FoodAIService {
 
         return foodPredictResponseDto;
     }
+
+    @Override
+    public List<FoodDetailInfoDto> getFoodDetailInfo(List<String> name, HttpServletRequest request) {
+        User user = jwtAuthenticationService.authenticationToken(request);
+
+        List<FoodNutrition> foodNutritionList = foodNutritionRepository.findFoodNutritionByFoodName(name);
+        log.info("[foodNutritionList] : {} ", foodNutritionList );
+        double expectedBloodSugar = calculateExpectedBloodSugar(user.getId(),foodNutritionList);
+        List<FoodDetailInfoDto> foodDetailInfoDtoList = foodNutritionList.stream().map(foodNutrition ->
+                new FoodDetailInfoDto(
+                        foodNutrition.getId(),
+                        foodNutrition.getName(),
+                        foodNutrition.getCarbohydrate(),
+                        foodNutrition.getProtein(),
+                        foodNutrition.getFat(),
+                        foodNutrition.getCholesterol(),
+                        foodNutrition.getGlIndex(),
+                        foodNutrition.getGiIndex(),
+                        expectedBloodSugar
+                )).collect(Collectors.toList());
+
+        request.getSession().setAttribute("foodDetailInfoDtoList", foodDetailInfoDtoList);
+        request.getSession().setAttribute("expectedBloodSugar", expectedBloodSugar);
+        return foodDetailInfoDtoList;
+    }
+
+    @Override
+    public ResultDto createAIMealInfo(String mealTime, HttpServletRequest request) {
+        User user = jwtAuthenticationService.authenticationToken(request);
+        List<FoodDetailInfoDto> foodDetailInfoDtoList = (List<FoodDetailInfoDto>)request.getSession().getAttribute("foodDetailInfoDtoList");
+        double expectedBloodSugar = (double)request.getSession().getAttribute("expectedBloodSugar");
+        String imageUrl = (String)request.getSession().getAttribute("imageUrl");
+        if (user == null) {
+            ResultDto resultDto = new ResultDto();
+            resultDto.setDetailMessage("사용자 인증에 실패했습니다.");
+            resultStatusService.setFail(resultDto);  // 실패 응답 설정
+            return resultDto;
+        }
+
+        // 음식 영양성분 아이디만 가져오기
+        List<Long> foodNutritionIds = foodDetailInfoDtoList.stream()
+                .map(FoodDetailInfoDto::getFoodId)
+                .collect(Collectors.toList());
+
+        // id로 영양성분 데이터 가져오기
+        List<FoodNutrition> foodNutritionList = foodNutritionRepository.findAllById(foodNutritionIds);
+
+        MealInfo mealInfo = MealInfo.MealInfo(mealTime,imageUrl,expectedBloodSugar,foodNutritionList,user);
+
+        mealDao.save(mealInfo);
+
+        // MealInfo 객체 생성
+        ResultDto resultDto = new ResultDto();
+        resultDto.setDetailMessage("식사 기록 저장 완료!");
+        resultStatusService.setSuccess(resultDto);
+
+        return resultDto;
+    }
+
+    public double calculateExpectedBloodSugar(Long userId, List<FoodNutrition> foodNutritionList) {
+        // 사용자 아이디로 최근 혈당 기록 가져오기
+        Optional<BloodSugar> currentBloodSugar = bloodSugarRepository.findClosestBloodSugar(userId);
+        log.info("[currentBloodSugar] : {} ", currentBloodSugar );
+        if (!currentBloodSugar.isPresent()) {
+            throw new RuntimeException("최근 등록된 혈당 기록이 없습니다.");
+        }
+
+        // 현재 혈당 값을 가져옴
+        double currentBloodSugarinfo = Double.parseDouble(currentBloodSugar.get().getBloodSugarLevel());
+        double totalBloodSugarIncrease = 0.0;
+
+        // 음식의 GI, GL, 탄수화물 정보를 기반으로 혈당 상승량 계산
+        for (FoodNutrition food : foodNutritionList) {
+            // GL 계산
+            double gl = (food.getGiIndex() * food.getCarbohydrate()) / 100.0;
+
+            // 예상 혈당 상승량 계산 (GL * 1.8)
+            double bloodSugarIncrease = gl * 1.8;
+
+            // 총 혈당 상승량에 더하기
+            totalBloodSugarIncrease += bloodSugarIncrease;
+        }
+        // 평균 혈당 상승량 계산
+        double averageBloodSugarIncrease = totalBloodSugarIncrease / foodNutritionList.size();
+
+        // 최종 예상 혈당 계산 (현재 혈당 + 평균 혈당 상승량)
+        double expectedBloodSugar = currentBloodSugarinfo + averageBloodSugarIncrease;
+
+        return expectedBloodSugar;
+    }
+
+
+
 }
+
